requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2==2.9.6
celery[redis]==5.3.1
flower==2.0.0
python-dotenv==1.0.0
requests